machine/qemu: Update the `State` method to be more efficient
This commit refactors QEMU's `State` method to be more efficient the
context of its usage.  When calling `State` for the QEMU driver, it will
use the information it has available to it at the time to more precisely
determine the state of the VMM whilst simultaneously updating the store
with the information it has determined to be correct and valid.  The
state is only saved if it now differs from what it has on record.

Signed-off-by: Alexander Jung <alex@unikraft.io>