unikraft/app: Disable use of `Set` and `Unset`
This commit disables the use of `Set` and `Unset`.  The purpose of this
change is that these two methods need to be re-thought towards their
implementation which should no longer be based on `DefConfig` and
instead use the existing configuration of the application to make the
appropriate adjustments.  Another open question is with regard to what
we are actually setting and where.  For example, does the use of these
methods simply change the contents of the Kraftfile or do they change
the values of the configuration one-time (e.g. useful for debugging or
quick feature swapping?)

Signed-off-by: Alexander Jung <alex@unikraft.io>