log: Introduce contextual logger
This commit introduces a new primitive to the `log` package which
enables access to a global logger via the use of the `context` package.
Its use is exported via the `G` variable to be used inline via:

```go
import (
  "context"
  "kraftkit.sh/log"
)

log.G(context.Background()).Info("example")
```

This will enable a refactor of the KraftKit codebase which will
eventually forgo and remove all instances of `WithLogger` option
methods.  Their use enabled a constraint which prevented logging without
access to a previously instantiated logger.  In this mode, we only need
to pass the richer `ctx` variable which, naturally, can be
contextualised, to use a specific logger.

Signed-off-by: Alexander Jung <alex@unikraft.io>