iostreams: Introduce contextual iostreams
This commit introduces a new primitive to the `iostreams` package which
enables access to a global iostreams via the use of the `context` package.
Its use is exported via the `G` variable to be used inline via:

```go
import (
"context"
"kraftkit.sh/iostreams"
)

iostreams.G(context.Background()).HasTrueColor()
```

This will enable a refactor of the KraftKit codebase which will eventually forgo
and remove all instances of `WithIOStreams` option methods or passing the
`IOStreams` structure directly.  Their use enabled a constraint which prevented
access to the iostream methods without access to a previously instantiated
`IOStreams` structure.  In this mode, we only need to pass the richer `ctx`
variable which, naturally, can be contextualised, to use a specific IO stream.

Signed-off-by: Alexander Jung <alex@unikraft.io>