config: Introduce contextual `ConfigManager`
This commit introduces a new primitive to the `config` package which
enables access to a global `Config` via the use of the `context` package.
Its use is exported via the `G` variable to be used inline via:

```go
import (
"context"
"kraftkit.sh/config"
)

// Where `ConfigOption` is an attirubte of `config.Config`
config.G(context.Background()).ConfigOption
```

This will enable a refactor of the KraftKit codebase which will eventually forgo
and remove all pass-by-reference paramaters which would otherwise always rely on
accessing user- or host-config for KraftKit.  Their use enabled a constraint
which prevented access to the config attribute without access to a previously
instantiated `Config` structure.  In this mode, we only need to pass the
richer `ctx` variable which, naturally, can be contextualised, to use a specific
IO stream.

Signed-off-by: Alexander Jung <alex@unikraft.io>