internal/cli: Register context and options in `New` method
This commit updates the `New` method such that it a). now receives
access to a context object and b). has passable options used to
customize the CLI, which itself initializes the context of the command.
This iterates through the available defaults and then applies all
user-specified options after which this context is passed to subcommands
and to the root command.

Signed-off-by: Alexander Jung <alex@unikraft.io>