Migrate to context-based logging
This commit applies the use of the context-based logger across the
project.  Its use is idealized via the following example:

```go
import "krafkit.sh/log"

log.G(ctx).Info("example")
```

This approach enables access to the logger with greater ease and
relieves the necessity to build any explicit pass-by-reference to a
logger which was always the case before this commit. Subsequently,
this commit removes all instances of option-driven methods, e.g.
`WithLogger`.

Additionally, to enable its use, the context package was imported and
applied as an ordered parameter to relevant methods.  This last change
is also necessary for additional context-based mechanisms used within
KraftKit and paves the way for its use in the future.  The largest
changes driven by context-based parameters occurs with the
`packagemanager.PackageManager` prototype.  This affects all adopting
implementations (notable the manifest and umbrella managers).  The
second largest change occurs in the `manifest` package to replace the
use of the pass-by-reference logger.

Signed-off-by: Alexander Jung <alex@unikraft.io>