Introduce centralised linting tool
This commit introduces a centralised linting tool[0] to be used in both
the CI as well as locally to check standard syntax issues in the
project.  The tool is more configurable, since it is an aggregation of
many, and therefore eases extensibility to new linters in the future.
An additional rule in the root Makefile is added to aid with checks
before submission as a PR.

[0]: https://golangci-lint.run/

Signed-off-by: Alexander Jung <alex@unikraft.io>