tui: Fix handling of `norender` option
This commit fixes both the `paraprogress` and `processtree` packages to
correctly handle the option `WithRenderer` such that the output is not
mangled by the tui itself.  Before this commit, the option was
ineffective.  Now with this option, and when set, each tui will no
longer affix the context with its bespoke `io.Writer` which enables the
specific view.  Such use is enabled in KraftKit later by setting the log
type to "fancy".

Signed-off-by: Alexander Jung <alex@unikraft.io>