cmd/kraft,internal/cli: Introduce subcommand grouping
This commit allows us to specify which group a subcommand belongs to
which allows for an easier-to-navigate help command output.  For now, 3
groups are provided: "build", "pkg" and "run" which display a set of
subcommands (and sub-sub-commands) which are part of the same group.  A
modificaiton to all current subcommand programs in KraftKit are made,
using this new system based on cobra's annotation system.

GitHub-Fixes: #36
Signed-off-by: Alexander Jung <alex@unikraft.io>