manifest: Do not use caching in manifest and index retrievals
This commit effectively reverts the caching mechanism introduced in
66ce21b.  Whilst the commit itself adds valid logic for calling HTTP
HEAD and correctly handling errors, the caching mechanism was shared
across both the `ManifestIndex` and `Manifest` structures.  This causes
an issue for two reasons: the cache is filled as soon as any request is
made and if either `NewManifestFromURL` or `MewManifestIndexFromURL` are
used in separate threads, the data can be corrupt since it does not have
a lock.

Signed-off-by: Alexander Jung <alex@unikraft.io>