manifest: Fix `CatalogQuery` attribute `Version` overwrite
This commit fixes an overwrite of the `Version` attribute in
`CatalogQuery` after the `Name` was parsed by the `GuessTypeNameVersion`
method.  The overwrite should only occur if the method returns a value.
In the case where it does not, the current value of `Version` should
remain.  This change ultimately fixes an issue where packages returned
from `Catalog` never had the right version if the `Version` attribute
was used as part of the query.

Signed-off-by: Alexander Jung <alex@unikraft.io>