Force just-in-time use of context with exec, make and `ApplicationConfig`
This commit applies the use of the first positional argument in the core
`ApplicationConfig` methods to use `ctx context.Context`.  The purpose
of this is 2-fold: 1). Many exposed interfaces in KraftKit are already
starting to use this type as the first positional argument and so it
homogenizes the codebase, and 2). the use of a `WithContext` option-
driven method applies a potentially out-of-date context -- this is the
case when using `ApplicationConfig` in the context of one of the KraftKit
TUIs.

Signed-off-by: Alexander Jung <alex@unikraft.io>