cmd/kraft/build,unikraft/app: Remove build option to run `syncconfig`
This commit removes the ability to run `make syncconfig` before the
build.  This is part of a larger issue and this represents a solve
around the larger issue which is to do with how stdin and stdout are
manipulated in the context of `kraft build`.  The `make syncconfig`
program becomes mangled and the build stalls as it is unable to receive
an input.  The workaround is to both remove this step and ensure that
`make defconfig` with the `kraft.yaml` KConfig values are always set.

Signed-off-by: Alexander Jung <alex@unikraft.io>