unikraft/app,cmd/kraft: Ensure all make-based methods use `TargetConfig`
This commit forces all main methods for interacting with Unikraft (and
therefore ones that eventually use the `ApplicationConfig.Make` method
to specify a `TargetConfig` as a positional parameter.  Omittion via
`nil` is possible however the API allows forces us to think about the
target in the context of the action performed by Unikraft.

Signed-off-by: Alexander Jung <alex@unikraft.io>