unikraft/app: Change the default target name to the standard kernel name
This commit changes the name of the target if it does not have a
pre-defined name such that it is in the format:

  `<PROJECT NAME>_<PLAT>-<ARCH>`

This is the built-in default binary name of the resulting unikernel.  It
allows us to eventually build per-target `.config` files which
accurately build the unikernel given the specific target and allows
clear separation between the target configurations.

Signed-off-by: Alexander Jung <alex@unikraft.io>