cmd/kraft,unikraft: Simplify project instantiation
This commit greatly simplifies the creation of a new `ApplicationConfig`
based on the "project option" mode.  In this instantiation method,
numerous "Kraftfiles" are read and parsed along with options as to which
checks and mechanisms are performed or along with additional data to
enrich the `ApplicationConfig`.

Much of the complexity was introduced by employing the loading mechanisms
from f2bb67c which imported code from the Docker Compose codebase.  The
design choice here split up both the loading and the instantiating of a
"project".  In this commit we essentially merge these facets since they
were greatly inter-dependent and would have not been used independently.
The result is a simpler API which is seen in all `cmd/*` changes.

Signed-off-by: Alexander Jung <alex@unikraft.io>