unikraft/app: Set the target kernel names in `NewApplicationFromOptions`
This commit fixes an issue where the target name was set inconsistently.
To mitigate this, we only set the kernel target name in the method
`NewApplicationFromOptions`.  This is called implicitly by the
`NewProjectFromOptions`.  Previously, an attempt to set the kernel name
was made by `LoadTarget`.  This function was unable to set the name
appropriately as we do not yet know the name by the time its called
during the parsing a Kraftfile.  Instead, standardise the `LoadTarget`
method to be in-line with other top-level attributes of a Kraftfile,
e.g. `LoadUnikraft` such that they have the same parameterisation.

After parsing all the Kraftfiles, `NewProjectFromOptions` performs a
"clean" instantiation based on all the information it has retrieved from
parsing all Kraftfiles.  As a result, we can now appropriately set the
name in `NewApplicationFromOptions` method.

Finally, we use the kernel name in the `ApplicationConfig` object's
`KConfigFile` method as opposed to the target name.  This provides us a
more distinct mechanism for identifying the which dot config resulted in
the final binary.

Signed-off-by: Alexander Jung <alex@unikraft.io>