kconfig,unikraft: Rename structures and methods to more appropriate names
This commit renames `KConfigValue` to `KeyValue` and `KConfigValues` to
`KeyValueMap` which are more appropriate representations of the
structures as well as clearer in code.  E.g.:

```go
kconfig.KConfigValue
// becomes
kconfig.KeyValue
```

The result propagates into the unikraft package.  At the same time, the
component methods for retrieving the KConfig key-value map is renamed to
`KConfig` and the method to access the menu is renamed to `KConfigTree`.

Signed-off-by: Alexander Jung <alex@unikraft.io>