unikraft: Migrate working directory/path-based attributes
This commit simplifies the use of the working directory as well as the
use of `ComponentConfig`.  Context and "component options" were used to
seed the `ComponentConfig` such that the working directory could be
determined (as well as each component's source directory).  This
complexity has been removed and added simply to the loader of each
component.

Signed-off-by: Alexander Jung <alex@unikraft.io>