unikraft/{lib,app}: Transform entire map
This commit updates the `LibraryConfig` transformer such that the whole
map is transformed as opposed to each individual item.  This allows us
to abstract the parsing of the schema into the library package whilst
simultaneously allowing us to populate the "name" field (which is
otherwise removed if only parsing the body from the map, since libraries
are key-value pairs and only the value is parsed).

Signed-off-by: Alexander Jung <alex@unikraft.io>