unikraft,cmd: Remove `ComponentConfig`
This commit migrates the attributes of `ComponentConfig` into the
individual components.  This abstraction was ultimately considered
unecessary and it complicated the use of each component.  This changeset
is quite large to adjust for this migration and includes the following
additional changes:

 - migrate the transformation function to each component, allowing the
   the component to be more compartmentalized in terms of related work
   (that is, only the component should know how to "unmarshal" itself
   from an interface);
 - introduce a general-purpose "Unikraft Context" which mimics variables
   that are part of the Unikraft core build system. For now these are
   only useful for a complete build.
 - pass context to each transformer such that we can access the Unikraft
   context which is wrapped in a general context structure.  Later on we
   can add additional things like logging or looking up information in
   the package manager thanks to this new access.
 - unexport (privatize) the attributes of each component and expose
   their value through public (and immutable) methods.

This reduction and simplification ultimately enables us to keep the
functionality of updating the attributes of the individual components
centralized to the component (that is, the value of each attribute is
highly controlled and should not necessarily be updated without proper
function).

Signed-off-by: Alexander Jung <alex@unikraft.io>