unikraft/core: Introduce `Libraries` component method
This commit introduces a new method which allows us to retrieve the list
of libraries which are provided by the Unikraft core.  It's achieved by
reading well-known paths in the core as well as the mechanisms
introduced in the library component which allow for reading in
arbitrary paths to be parsed as a library.

Signed-off-by: Alexander Jung <alex@unikraft.io>