Refactor packaging interface
This broad commit creates clarity between subsystems which implement the
`packmanager.PackageManager` and `pack.Package` interface.  Now there is
clear separation between the two, where the `PackageManager` now is used
to:

 1. Add and remove sources which return lists of packages;
 2. Allow for querying sources to find specific packages; and,
 3. Packaging (`Pack`ing) a component into an implementation-specific
    package and unpackaging (`Unpack`ing) a provided package into a
    component.

The last destinction (3.) was somewhat unclear in the previous
implementation of KraftKit which this commit resolves.  Specifically,
the packaging of a `unikraft.component.Component` type was handled in
`cmd/kraft/pkg/pkg.go` when it should simply call the package manager
with the provided project (which implements the `Component` interface).
This allows the implementing interface to handle different component
types and allows the APIs of KraftKit be used to package components
which was not previously possible.

Since the `PackageManager` interface resolves `Package`s, the operation
of `Pull`ing and `Push`ing have now been introduced as interface methods
to `pack.Package`.  This is because the implementing package will have
contextual information about where its origin is from and so it is
possible to retrieve (`Pull`) from this origin as well as `Push` to it.

Signed-off-by: Alexander Jung <alex@unikraft.io>