unikraft: Struct to describe "provided syscalls"
In the Unikraft core, a microlibrary/component can expose syscalls which
it registers in a number of ways.  In KraftKit we can build numerous
ways for recognising these syscalls which aid in a high-level
understanding and parsing of the library (e.g. to determine
dependencies).  However, to centralize the idea of a syscall, we add a
`unikraft.ProvidedSyscall` struct which can be used to hold the name and
number of positional arguments.  We can use this structure later with
`LibraryConfig` to list its provided syscalls.

Signed-off-by: Alexander Jung <alex@unikraft.io>