unikraft/app,cmd/kraft: Unexport the application structure
This commit unexports the "application struct" which implements the
`Application` interface.  There is no need to have a separately
addressable symbol which can also not be manipulated since its
attributes are private and whose methods implement the already
ubiquitious `Application` interface.  This "privatization" ensures
strong internal manipulation of the structure and only exposes
useful and well-defined methods via the corresponding interface.

Signed-off-by: Alexander Jung <alex@unikraft.io>