Statically link libgit2-dev
This commit statically links binaries of KraftKit with libgit2.  To do
this we vendor the git2go package which requires libgit2.  However, to
prevent vendoring every required Go module, we a). include git2go as a
Git submodule and b). explicitly ignore the vendor directory in the `go
build` command by passing `-mod=readonly` flag.

This commit simultaneously ignores the vendor directory in the CI
configuration and makes cloning the libgit2 repository a dependency for
when starting the developer environment which would otherwise result in
a dubious ownership of the submodule if cloned with a different user and
group ID.

GitHub-Fixes: #179
GitHub-Closes: #31
Signed-off-by: Alexander Jung <alex@unikraft.io>