.vscode, .gitignore: Include VS Code settings
This commit includes workspace settings for VS Code, particularly
configuration which manipulates the VS Code Extension for Go.  Due to
partial vendoring of the project, we must signify to Go to use
`-mod=readonly` in both the build process (if executed through VS
Code itself) and to the Go language service which would otherwise error
and prevent any in-IDE utilities such as code-completion.

Signed-off-by: Alexander Jung <alex@unikraft.io>