Initialize go workspace
This commit introduces a new top-level file `go.work`, an axuiliary file
made available from Go 1.18 which signifies a repository with multiple
Go modules.  Since there are essentially 3 modules within KraftKit (so
far) we introduce this file to aid the Go compiler and Go tools
navigating the repository.

In addition to this, given that we are now partially vendoring the
repository due to the inclusion of libgit2, this file helps in the
vendoring-realization process for Go tools such as with `gopls`.

Signed-off-by: Alexander Jung <alex@unikraft.io>