Set the static tag in the GoReleaser build
With partial vendoring of libgit2, we must include the `static` tag in
any `go build` invocation such that we use the locally vendored shared
object file.  This commit updates GoReleaser's configuration to set this
tag and to signify this tag to `go build` such that the relevant
tag-specific Go files which link against the shared object are indeed
made.

Signed-off-by: Alexander Jung <alex@unikraft.io>