unikraft/app: Pin ordering for libC's
Because libraries are kept in a map, their ordering can't be guaranteed
This comes in conflict with libC-type libraries, which need to sit
first. This is not a problem related with them, and comes from other
libraries using the `#include_next` syntax on them.

To fix this we pin either `musl` or `newlib` plus `pthread-embedded` to
come first, if they exist.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>