cmd,config: Convert `ConfigManager` into a generic
This broad commit converts the use of the globally accessible
configuration, typically accessed via `config.G(ctx)` into a generic
such that the use of the `kraftkit.sh/config` is abstract to KraftKit.
Whilst KraftKit configuration is still embedded, now at
`config.KraftKit`, this is not enforced in the package's instantiation.
This is enforced only when the manager is invoked and the initial type
is provided.

Signed-off-by: Alexander Jung <alex@unikraft.io>