tools/webinstall: Add install script
The shell script is the central piece of the tool. It contains all the
steps needed to install kraftkit through package managers or directly
from GitHub.

The script is written to be compliant with POSIX-sh, which should
guarantee it works on any distribution and shell combination.

For Windows the script will not run, unless it is ran on Cygwin.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>