tools/webinstall: Add minimalistic page server
The Go Server is minimalistic and offers only two endpoints, one that
serves the install script and one that offers the latest version of
kraftkit. The server can be configured how often it should query
GitHub, the port to serve data on, and a GitHub token to avoid
rate limiting.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>