fix(run): correctly handle executing and tailing logs of a unikernel
This commit fixes the instantiation of a unikernel, tailing its logs and
preventing the execution of the spawned process (`kraft events`) which
did not correctly handle the facilitation of unikernel exits.

First, the instantiation of the events monitor is removed from `kraft
run`.  The original purpose of this hidden subcommand was to aid in
removing or destroying of a unikernel.  Instead this functionality is
added directly to `kraft run`.

Secondly, the actual instantiation of the unikernel via `driver.Start()`
is moved as late-as-possible such that go-routines can be established
first.  The removal or destruction of a unikernel is affixed at the end
of the execution such that we can wait for an exit event.  Having the
event within the context of a Go routine would prematurely exit the
program and thus the unikernel could not be stopped or removed.

Finally, manipulation of the cancelable-context is fixed with
respect to the fixed `TailWriter`.

Signed-off-by: Alexander Jung <alex@unikraft.io>