fix(lib): Always set the library's KConfig option
This commit fixes setting the library's KConfig value which would
otherwise not be set if the library's Config.uk file could not be
parsed.  The KConfig parser is currently broken for Unikraft
microlibraries which do not generally wrap `menuconfig` with
`endmenu`.  Additional changes are required such that this is
recognised for each microlibrary by the parser OR a fix occurs in
each microlibrary up stream (which is unlikely).  Either way, if
the menu cannot be read, naively set the KConfig name for this
library.

Signed-off-by: Alexander Jung <alex@unikraft.io>