fix(tui): Handle non-interactive environments
This fixes an issue where we do not wish to render an input TTY
into the TUI since it is a pseudo environment used to simply output
logs.  As a result, we wish to pass `nil` as stdin to bubbletea.  Due to
a separate issue [0], we have to simultaneously upgrade bubbletea.

[0]: <https://github.com/charmbracelet/bubbletea/pull/643>

Signed-off-by: Alexander Jung <alex@unikraft.io>