feat(iostreams): Upgrade iostreams for better pseudo TTY handling
This commit updates the iostreams package based on upstream changes from
its original source [0].  These reduce iostreams complexity and resolve
a number of bugs.  More importantly, it actualizes the
`IOStreams.{In,Out}` attributes to a more "file-like" interface such
that they return `Fd`. This can be later used to illustrate whether
stdout and stdin are from a real or pseudo TTY source.

[0]: <https://github.com/cli/cli/commits/trunk/pkg/iostreams/iostreams.go>

Signed-off-by: Alexander Jung <alex@unikraft.io>