fix(kraft): Fix a race condition between checking update with config
This commit fixes a race condition with KraftKit's use of contexts.  The
use of `Pre` is not possible with the root command because the context
is populated in `Run` and so invocations and access to the provided
`ctx` returns unfufilled values.  In this case, an attempt to retrieve
the latest version of KraftKit is made in the `Pre` method.  Since this
invocation relies on the `ConfigManager` to be initialized, it has
insufficient knowledge to continue.  The result of invoking the
`ConfigManager` via `ctx` before its instantiation is that default
values would not be properly fufilled.  Following commands, such as
`kraft pkg update` would be improperly set.

This commit therefore moves the check for an update after the
instantiation of the general context for the main programs invocation.

Signed-off-by: Alexander Jung <alex@unikraft.io>