refactor!: Introduce package manager constructor
This commit introduces a constructor prototype that should be used for
all implementing package managers.  This standardizes the instantiation
of a new package manager such that its registration (and later
initialization) accept the same parameters.

The constructor enables access to the globally initialized context.  As
a result of requiring the context, the "umbrella package manager" needs
to be initialized after full context has been made.  As a result, the
CLI option `WithDefaultPackageManager` is removed and the PM is
instantiated after the config manager, logger and IOStreams is ready.

Signed-off-by: Alexander Jung <alex@unikraft.io>