docs(internal): Squash the license and copyright notice preamble
This commit "squashes" the license and copyright preamble in files
associated with this sign off.  The preamble already includes the SPDX
license identifier which essentially duplicates the text in question.
Additionally, this license, in accordance with the license, is
distributed with the source code via the `LICENSE.md` file which is
accompanied in this repository.

Additionally, copyright notice is provided to The KraftKit Authors,
which entitles new developers who contribute commits to the project an
associated copyright with regard to their contributions without having
to specifically update the license and copyright preamble.

Signed-off-by: Alexander Jung <alex@unikraft.io>