fix(packmanager): Do not error-out when initializing package manager
This commit prevents an error from occurring when the package manager
cannot be initialized via its constructor.  Instead, a warning message
is shown instead.  This allows, for example, the progression of
applications which may not necessarily rely on a package manager to
continue without erroring if there is an issue instantiating 1 or more.

Signed-off-by: Alexander Jung <alex@unikraft.io>