refactor!: Remove `ArchPlatString` in favor of `TargetPlatArchName`
Not only is the order of "Platform" and "Architecture" corrected in the
rename, but the refactor simplifies any implementation of the `Target`
interface since the result of this function would always be the same;
that is, that the architecture and platform are always concatenated in
the same way (e.g. `kvm-x86_64`) since this is a Unikraft naming
convention.

Signed-off-by: Alexander Jung <alex@unikraft.io>