feat(run): Attempt package catalog twice: first local then remote
This commit enables querying the catalog first for a local reference of
the package which may already exist for the user and who wishes to run
instead of always forcing the remote use of a catalog.  If, however, the
package cannot be found locally, a request to update the catalog via
`NoCache` option is set when it is not available locally.

Signed-off-by: Alexander Jung <alex@unikraft.io>