fix(manifest): Re-order the provider detection, prioritize directories
This commit re-orders the detection of an input manifest provider such
that the directory provider has higher priority (2nd, after index native
YAML representations) and above Git and GitHub.  This is because a user
may wish to hack or use a local directory representing a Unikraft
microlibrary.  The order is also above Git, because the Git provider
would parse the `.git/` folder and return additional metadata
information.  If the user is providing a directory, more than likely
they wish to use it without the metadata since it is not remote.

Signed-off-by: Alexander Jung <alex@unikraft.io>