fix(oci): Correctly handle full reference names
This commit fixes two instances where a full reference to the image is
required in the format `name:tag` which are required when packaging and
fetching images locally.

Without this change, the `latest` tag is used by default which may not
be the case.  To accommodate the change a new standard `imageRef`
method is introduced to return this string value.

Signed-off-by: Alexander Jung <alex@unikraft.io>