refactor(qemu): Prefix qemu-system-* structures/constants
This commit prefixes all files related to direct interfaces with the
qemu-system-* binaries (and therefore offer the Go abstractions which
map to the CLI flags) with qemu_.  This ensures that files related to
this part of the package vs. other parts (such as configuration,
machine driver, etc.) are clearly separated.

Signed-off-by: Alexander Jung <alex@unikraft.io>