feat(api): Introduce top-level API package
This commit introduces a top-level API package which exposes KraftKit
internal APIs via the Zip API Object Framework[0].  The framework allows
the definition of APIs in a Kubernetes-like fashion without requiring
code generation.  Therefore it is possible to generate traditional
CRD-like objects (objects usually defined with a "Spec" and "Status") in
a simple way.

The use of the Zip API Object Framework additionally provides a
configurable client which is protocol agnostic, allowing the interface
to the object to be defined by the implementer.  One potential client
implementation could simply be an inline function call, a remote HTTP
request, gRPC call, etc.

This initial commit introduces the definition of the "Machine Interface"
for KraftKit, which essentially defines the configuration used within
the context of the current machine driver.  Whilst this API is not used
within KraftKit as of this commit, this provides a base towards how the
driver interface can exist in the future: independent, well-defined,
versioned.

[0]: https://api.zip

Signed-off-by: Alexander Jung <alex@unikraft.io>