feat(set): Introduce additional set helper methods
This commit creates a distinction between finding any or exactly 1 item
in a set.  Before, the use of the string set `Contains` would return if
any items were found.  This may have unintended effect when we are
expecting to find exact match.

Necessary adjustments of the new methods are made in the project.

Signed-off-by: Alexander Jung <alex@unikraft.io>