refactor!: Use the API-defined machine interface
This large refactor deprecates the `machine/driver` package in favour of
the new top-level API package.  To make this switch, the qemu driver
implementation has been adjusted to match the service implementation
defined in `api/machine/v1alpha1`.

All uses of driver in subcommands are adjusted simultaneously to keep
the project in working condition.

Part of this refactor adjusts how the `kraft run` command works due to
the nature of the change of the API. Particularly, the split between
"kernel arguments" and "application arguments" is made more prevalent
through the API.  As such, the args supplied to `kraft run` has changed,
which now has an additional `--kernel-args` flag.

Signed-off-by: Alexander Jung <alex@unikraft.io>