fix(plat): Handle KVM VMM target names
Following the merge of [0] and the release of Unikraft v0.13.0 (Atlas)
comes the introduction of KVM VMM targets.  This has the effect of
modifying the resulting kernel binary name with the VMM of choice.
This commit allows for specifying this target and correctly handling
setting the correct internal KConfig options.

[0]: https://github.com/unikraft/unikraft/pull/760

Signed-off-by: Alexander Jung <alex@unikraft.io>