refactor!: Use option-methods for Catalog and IsCompatible
This commit refactors the prototypes for the Package Manager's methods
`Catalog` and `IsCompatible` such that

a). both methods now receive the underlying values of the query;
b). are using the option-method pattern which aligns with the rest of
    KraftKit's convention for method parameterization; and,
c). introduces the "auth" PullOption to enable per-pull authentication.

For a), passing these values will later prove valuable to the
`IsCompatible` method and its underlying implementations to help
determine whether the provided set of parameters are compatible for
the given package manager.

Signed-off-by: Alexander Jung <alex@unikraft.io>