feat: Add ability to expose port
This commit makes a cross-system change which demonstrates and enables
the ability to add a new API interfaces and relevant changes to the
relevant subsystems whilst simultaneously adding support for exposing a
a port to a unikernel run via QEMU via the `kraft run -p` syntax.  The
CLI syntax is identical to other container runtime CLI tools like
`docker` so as to mimic their use and make `kraft` similar in feeling.

Additionally, the `kraft ps` subcommand is adjusted to show the port of
a running unikernel instance.

Signed-off-by: Alexander Jung <alex@unikraft.io>