fix(handler): Fix calculation of progress based on host active tasks
In the event that containerd has on-going tasks that are not relevant to
KraftKit and the OCI handler for containerd, this progress calculation
would linger continuously (or until those statuses are ultimately
resolved) therefore representing an inaccurate progress.  If a status
was paused externally, the handler would also linger forever, resulting
in hung progress statuses reported by the handler.

Adjust the `ListStatuses` method to include a filter and check these
per job sequence calculation.

Signed-off-by: Alexander Jung <alex@unikraft.io>