refactor(run): Introduce runner-led entrypoints for `kraft run`
This commit refactors the `kraft run` subcommand to be more extensible
and to clearly separate the concerns by which the first positional
argument of `kraft run` is used to determine _what_ to run.

In this refactor, the 3 main ways in which a unikernel be run are
clearly separated, both in terms of individual files, but also through
the new abstract interface `runner` which requires implementation to
both determine if possible and then later (if possible) to prepare the
machine interface for the instantiation (population of specification) of
the unikernel VM.  Necessary adjustments to the preparation
(`cmdfactory`'s `Pre` method) and overall body of `cmdfactory`'s `Run`
interface are adjusted, resulting in a larger diff.

The following 3 methods are capable (currently) through the `kraft run`
subcommand:

1. kernel: The first positional argument represents a path to a Unikraft
   unikernel, where the user knows the path of the kernel following a
   build;
2. project: The first positional argument or zero-positional argument
   and current working directory are determined to be a KraftKit-enabled
   repository path (one with a `kraft.yaml`/`Kraftfile` schema); and,
3. package: The first positional argument represents a
   `PackageManager`-compatible package which contains a pre-built
   unikernel.

Signed-off-by: Alexander Jung <alex@unikraft.io>